# expression: implement vectorized evaluation for `builtinSetVarSig` and `builtinGetVarSig`