--- a/expression/builtin_other_vec.go
+++ b/expression/builtin_other_vec.go
@@ -14,8 +14,12 @@
 package expression
 
 import (
+	"strings"
+
 	"github.com/pingcap/errors"
+	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util/chunk"
+	"github.com/pingcap/tidb/util/stringutil"
 )
 
 func (b *builtinValuesIntSig) vectorized() bool {
@@ -91,11 +95,42 @@ func (b *builtinGetParamStringSig) vecEvalString(input *chunk.Chunk, result *chu
 }
 
 func (b *builtinSetVarSig) vectorized() bool {
-	return false
+	return true
 }
 
 func (b *builtinSetVarSig) vecEvalString(input *chunk.Chunk, result *chunk.Column) error {
-	return errors.Errorf("not implemented")
+	n := input.NumRows()
+	buf0, err := b.bufAllocator.get(types.ETString, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(buf0)
+	if err := b.args[0].VecEvalString(b.ctx, input, buf0); err != nil {
+		return err
+	}
+	buf1, err := b.bufAllocator.get(types.ETString, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(buf1)
+	if err := b.args[1].VecEvalString(b.ctx, input, buf1); err != nil {
+		return err
+	}
+	result.ReserveString(n)
+	sessionVars := b.ctx.GetSessionVars()
+	sessionVars.UsersLock.Lock()
+	defer sessionVars.UsersLock.Unlock()
+	for i := 0; i < n; i++ {
+		if buf0.IsNull(i) || buf1.IsNull(i) {
+			result.AppendNull()
+			continue
+		}
+		varName := strings.ToLower(buf0.GetString(i))
+		res := buf1.GetString(i)
+		sessionVars.Users[varName] = stringutil.Copy(res)
+		result.AppendString(res)
+	}
+	return nil
 }
 
 func (b *builtinValuesDecimalSig) vectorized() bool {
@@ -107,9 +142,34 @@ func (b *builtinValuesDecimalSig) vecEvalDecimal(input *chunk.Chunk, result *chu
 }
 
 func (b *builtinGetVarSig) vectorized() bool {
-	return false
+	return true
 }
 
 func (b *builtinGetVarSig) vecEvalString(input *chunk.Chunk, result *chunk.Column) error {
-	return errors.Errorf("not implemented")
+	n := input.NumRows()
+	buf0, err := b.bufAllocator.get(types.ETString, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(buf0)
+	if err := b.args[0].VecEvalString(b.ctx, input, buf0); err != nil {
+		return err
+	}
+	result.ReserveString(n)
+	sessionVars := b.ctx.GetSessionVars()
+	sessionVars.UsersLock.Lock()
+	defer sessionVars.UsersLock.Unlock()
+	for i := 0; i < n; i++ {
+		if buf0.IsNull(i) {
+			result.AppendNull()
+			continue
+		}
+		varName := strings.ToLower(buf0.GetString(i))
+		if v, ok := sessionVars.Users[varName]; ok {
+			result.AppendString(v)
+			continue
+		}
+		result.AppendNull()
+	}
+	return nil
 }