PCP #12106
<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
implement vectorized evaluation for  `builtinSetVarSig` and `builtinGetVarSig` , for #12106

### What is changed and how it works?
```
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinOtherFunc/builtinSetVarSig-VecBuiltinFunc-3         	    5208	    236473 ns/op
BenchmarkVectorizedBuiltinOtherFunc/builtinSetVarSig-NonVecBuiltinFunc-3      	    3740	    301244 ns/op
BenchmarkVectorizedBuiltinOtherFunc/builtinGetVarSig-VecBuiltinFunc-3         	    6942	    189947 ns/op
BenchmarkVectorizedBuiltinOtherFunc/builtinGetVarSig-NonVecBuiltinFunc-3      	    4358	    272404 ns/op
PASS
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test